Fetch measurements with redux
