+++ src/reducers/sensorReducer.js
@@ -2,6 +2,8 @@ const INITIAL_STATE = {
    sensors: [],
    error: null,
    isLoading: true,
+   measurements: [],
+   isMeasurementsLoading: true,
 }
 
 export default (state = INITIAL_STATE, action) => {
@@ -12,6 +14,12 @@ export default (state = INITIAL_STATE, action) => {
       return { ...state, sensors: action.payload, isLoading: false}
     case "GET_SENSORS_ERROR":
       return { ...state, error: action.payload, isLoading: false }
+    case "FETCH_MEASUREMENTS":
+      return { ...state, isMeasurementsLoading: true } 
+    case "FETCH_MEASUREMENTS_SUCCESS":
+      return { ...state, measurements: action.payload, isMeasurementsLoading: false }
+    case "FETCH_MEASUREMENTS_ERROR":
+      return { ...state, error: action.payload, isMeasurementsLoading: false }
     default:
       return state
   }
